// src/interfaces/licencia.interface.ts
export interface LicenciaFuncionamiento {
    uuid?: string;
    propietario_nombre?: string;
    rfc?: string;
    numero_patente?: string;
    giro_comercial?: string;
    giro_secundario?: string;
    tipo_operacion?: string;
    tipo_zona?: string;
    horario_operacion?: string;
    domicilio_fiscal?: string;
    fecha_impresion?: string;
    observaciones?: string;
}   